// packages/puppeteer-extra-plugin-recaptcha/src/types.ts
export type CaptchaVendor = 'recaptcha' | 'hcaptcha';

export type CaptchaType = 'checkbox' | 'invisible' | 'score';

export interface CaptchaInfo {
  _vendor: CaptchaVendor;
  id?: string; // captcha id
  widgetId?: number;
  sitekey?: string;
  s?: string; // new google site specific property
  isEnterprise?: boolean;
  isInViewport?: boolean;
  /** Is captcha invisible */
  isInvisible?: boolean;
  /** Invisible recaptchas: Does the captcha have an active challenge popup */
  hasActiveChallengePopup?: boolean;
  /** Invisible recaptchas: Can the captcha trigger a challenge or is it purely score based (v3) */
  hasChallengeFrame?: boolean;
  _type?: CaptchaType;
  action?: string; // Optional action (v3/enterprise): https://developers.google.com/recaptcha/docs/v3#actions
  callback?: string | Function;
  hasResponseElement?: boolean;
  url?: string;
  display?: {
    size?: string;
    theme?: string;
    top?: string;
    left?: string;
    width?: string;
    height?: string;
  };
}

export type FilteredCaptcha = CaptchaInfo & {
  filtered: boolean;
  filteredReason:
    | 'solveInViewportOnly'
    | 'solveScoreBased'
    | 'solveInactiveChallenges';
};

export interface CaptchaSolution {
  _vendor: CaptchaVendor;
  id?: string; // captcha id
  provider?: string;
  providerCaptchaId?: string;
  text?: string; // the solution
  requestAt?: Date;
  responseAt?: Date;
  duration?: number;
  error?: string | Error;
  hasSolution?: boolean;
}

export interface CaptchaSolved {
  _vendor: CaptchaVendor;
  id?: string; // captcha id
  responseElement?: boolean;
  responseCallback?: boolean;
  solvedAt?: Date;
  error?: string | Error;
  isSolved?: boolean;
}
// ...
export interface ContentScriptOpts {
  visualFeedback: boolean;
  debugBinding?: string;
}

export interface ContentScriptData {
  solutions?: CaptchaSolution[];
}